requirements.txt
<<<<<<< HEAD
requests==2.31.0
beautifulsoup4==4.12.3
Pillow==10.2.0
=======
scrapy==2.11.0
pillow==10.2.0
requests==2.31.0
beautifulsoup4==4.12.3
aiohttp==3.10.5
python-dotenv==1.0.0
scrapy-splash==0.9.0
scrapy-selenium==0.0.7
selenium==4.11.2
webdriver-manager==4.0.1
>>>>>>> 37aa198425ea01b51bbf147264fe11b1444653d6
